Add replace outdated example with a simpler one and add gitignore
